+++ src/components/bytes/Edit/EditByteStepperItem.tsx
@@ -23,7 +23,7 @@ import isEqual from 'lodash/isEqual';
 import { useState } from 'react';
 import styled from 'styled-components';
 import { v4 as uuidv4 } from 'uuid';
-import { predefinedColors } from '@/utils/colors/findClosestColor';
+import { findClosestColor } from '@/utils/colors/findClosestColor';
 
 interface EditByteStepperItemProps {
   space: SpaceWithIntegrationsFragment;
@@ -67,34 +67,6 @@ export default function EditByteStepperItem({
   updateStep,
 }: EditByteStepperItemProps) {
   const [modalByteInputOrQuestionOpen, setModalByteInputOrQuestionOpen] = useState(false);
-
-  function hexToRgb(hex: string): [number, number, number] {
-    const r = parseInt(hex.slice(1, 3), 16);
-    const g = parseInt(hex.slice(3, 5), 16);
-    const b = parseInt(hex.slice(5, 7), 16);
-    return [r, g, b];
-  }
-
-  function colorDistance(color1: [number, number, number], color2: [number, number, number]): number {
-    return Math.sqrt(Math.pow(color1[0] - color2[0], 2) + Math.pow(color1[1] - color2[1], 2) + Math.pow(color1[2] - color2[2], 2));
-  }
-
-  function findClosestColor(hex: string): string {
-    const rgb = hexToRgb(hex);
-    let closestColor: Color | null = null;
-    let smallestDistance = Number.MAX_VALUE;
-
-    predefinedColors.forEach((predefinedColor) => {
-      const predefinedRgb = hexToRgb(predefinedColor.hex);
-      const distance = colorDistance(rgb, predefinedRgb);
-      if (distance < smallestDistance) {
-        smallestDistance = distance;
-        closestColor = predefinedColor;
-      }
-    });
-
-    return closestColor ? (closestColor as Color).name : 'Unknown';
-  }
 
   const updateStepContent = (content: string) => {
     updateStep({ ...step, content });
